Modified births process so that HIV+ children are drawn binomially rather than as floor(p*n) since p*n<1.

+++ IBM_simul/src/demographics.h
@@ -26,7 +26,7 @@ double childhood_mortality(parameters *, double );
 double natural_death_rate(int , int , parameters *, double );
 int draw_sex_risk(int, parameters *);
 void create_mtct_templates(mtct_hiv_template *, parameters *);
-int get_art_status_of_mtct_new_adult(int , parameters *);
+int get_art_status_of_mtct_new_adult(double , parameters *);
 void add_hiv_info_for_new_hiv_positive_adult(individual *, int, double , parameters *, patch_struct *, int );
 void create_new_individual(individual *, double , parameters *, int, 
     patch_struct *, int, all_partnerships *);
